+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## master
+
+* Add `MacAddress` field to `ServerPrivateNet`
+
 ## v1.14.0
 
 * Add `Created` field to `FloatingIP`